requirements.txt
pyannote-audio==3.1.1